Updates to ComplexPlane.cpp

+++ ComplexPlane.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <SFML/Graphics.hpp>
 #include <iostream>
+#include <complex>
 
 using namespace std;
 using namespace sf;
@@ -33,7 +34,7 @@ private:
 	int m_zoomCount;
 	float m_aspectRatio;
 
-	int countIterations(Vector2f coord);
+	size_t countIterations(Vector2f coord);
 	void iterationsToRGB(size_t count, Uint8& r, Uint8& g, Uint8& b);
 	Vector2f mapPixelToCoords(Vector2i mousePixel);
 };
